requirements.txt
boto3==1.34.154
botocore==1.34.154
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
greenlet==3.0.1
idna==3.7
jmespath==1.0.1
joblib==1.4.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
numexpr==2.10.1
numpy==1.26.4
pandas==2.2.2
pip==24.2
psycopg2==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.1
scikit-learn==1.5.1
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.34
tabula-py==2.6.0
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.3
wheel==0.44.0
win-inet-pton==1.1.0
